Moduleify MathDetails component
